Move active after focus in example
If `focus` is defined after `active`, then those styles will take precedence over the `active` styles, leading to the `active` styles to never be used if there is a conflict (I believe anytime something is active it is also focused).

This is of course just an example and the end user can do whatever they want. But using the docs as an assumed decent starting point, it took me a few minutes to realize what was happening with my active styles.